Release v3.2

+++ src/radio/radio.h
@@ -226,16 +226,16 @@ struct Radio_s
      */
     bool    ( *CheckRfFrequency )( uint32_t frequency );
     /*!
-     * \brief Computes the packet time on air for the given payload
+     * \brief Computes the packet time on air in us for the given payload
      *
      * \Remark Can only be called once SetRxConfig or SetTxConfig have been called
      *
      * \param [IN] modem      Radio modem to be used [0: FSK, 1: LoRa]
      * \param [IN] pktLen     Packet payload length
      *
-     * \retval airTime        Computed airTime for the given packet payload length
+     * \retval airTime        Computed airTime (us) for the given packet payload length
      */
-    double  ( *TimeOnAir )( RadioModems_t modem, uint8_t pktLen );
+    uint32_t  ( *TimeOnAir )( RadioModems_t modem, uint8_t pktLen );
     /*!
      * \brief Sends the buffer of size. Prepares the packet to be sent and sets
      *        the radio in transmission
